Refactor the structure tests a bit. This will let us reuse the parsing/types for other code.
In particular, we'll be able to add a "schema validator".
